+++ internal/parser/utils.go
@@ -3,7 +3,7 @@ package parser
 import "strconv"
 
 // parseInt transforms a string into an int.
-func parseInt(in string) int64 {
+func parseInt(in string) int {
 	v, _ := strconv.ParseInt(in, 10, 64)
-	return v
+	return int(v)
 }
